Update PowerShell SKU management - wip
The original purpose was to bulk unassign and reassign the E3 licenses on a tenant where when the e3 license was originally purchased EXO plan 2 limit was 50 gb and not as it is now (2/5/2020) 100 gb. Internal diagnostics did not correct the service plan failure but unassigning and reassigning the license did hence this script to bulk edit licenses.

Future plans for this script including building a gui and making it so it is much more general use.